Fixed bug with camera lerp not working all the time. Improved windzone and now soundevent uses byte as Strenght value instead of integer

+++ Assets/Scripts/Audio/Reactive/AudioReactiveClip.cs
@@ -6,7 +6,7 @@ using System.Linq;
 public struct ReactiveClipData
 {
     public float Timestamp;
-    public int Strenght;
+    public byte Strenght;
 }
 [CreateAssetMenu(menuName = "Audio/ReactiveClip")]
 public class AudioReactiveClip : ScriptableObject
